// src/Types/ColorEnum.ts
enum ColorEnum {
    BLACK = '%black%',
    DARK_RED = '%dark_red%',
    DARK_GREEN = '%dark_green%',
    GOLD = '%gold%',
    DARK_BLUE = '%dark_blue%',
    VIOLET = '%violet%',
    CYAN = '%cyan%',
    GREY = '%grey%',
    DARK_GREY = '%dark_grey%',
    RED = '%red%',
    GREEN = '%green%',
    YELLOW = '%yellow%',
    BLUE = '%blue%',
    PINK = '%pink%',
    LIGHT_BLUE = '%light_blue%',
    WHITE = '%white%'
}

export default ColorEnum;